fix: remove strong reference retain cycle from camera preview view model
