+++ Sources/FaceLiveness/Views/GetReadyPage/CameraPreviewViewModel.swift
@@ -27,9 +27,13 @@ class CameraPreviewViewModel: NSObject, ObservableObject {
             position: .front
         ).devices.first
 
+        let outputDelegate = CameraPreviewOutputSampleBufferDelegate { [weak self] buffer in
+            self?.updateBuffer(buffer)
+        }
+
         self.previewCaptureSession = LivenessCaptureSession(
             captureDevice: .init(avCaptureDevice: avCaptureDevice),
-            outputDelegate: self
+            outputDelegate: outputDelegate
         )
         
         do {
@@ -52,18 +56,10 @@ class CameraPreviewViewModel: NSObject, ObservableObject {
     func stopSession() {
         previewCaptureSession?.stopRunning()
     }
-}
 
-extension CameraPreviewViewModel: AVCaptureVideoDataOutputSampleBufferDelegate {
-    func captureOutput(
-        _ output: AVCaptureOutput,
-        didOutput sampleBuffer: CMSampleBuffer,
-        from connection: AVCaptureConnection
-    ) {
-        if let buffer = sampleBuffer.imageBuffer {
-            DispatchQueue.main.async {
-                self.buffer = buffer
-            }
+    func updateBuffer(_ buffer: CVImageBuffer) {
+        DispatchQueue.main.async {
+            self.buffer = buffer
         }
     }
 }
